Broke out race unsafe into separate file to allow race checking during tests. Also fixed location of tmp files to run under os.Tmp for use with networked storage
