+++ treap.go
@@ -292,12 +292,17 @@ func (o *Store) walk(t *Collection, withValue bool, cfn func(*node) (*nodeLoc, b
 func (o *Store) visitNodes(t *Collection, n *nodeLoc, target []byte,
 	withValue bool, visitor ItemVisitorEx, depth uint64,
 	choiceFunc func(int, *node) (bool, *nodeLoc, *nodeLoc)) (bool, error) {
+	save_mem := false
+  save_mem = true
 	nNode, err := n.read(o)
 	if err != nil {
 		return false, err
 	}
 	if n.isEmpty() || nNode == nil {
 		return true, nil
+	}
+	if save_mem {
+		//fer nNode.Evict()
 	}
 	nItemLoc := &nNode.item
 	nItem, err := nItemLoc.read(t, false)
@@ -307,8 +312,6 @@ func (o *Store) visitNodes(t *Collection, n *nodeLoc, target []byte,
 	if nItem == nil {
 		panic(fmt.Sprintf("visitNodes nItem nil: %#v", nNode))
 	}
-
-	save_mem := false
 
 	choice, choiceT, choiceF := choiceFunc(t.compare(target, nItem.Key), nNode)
